refactor: data 폴더, ui 폴더 분리
