+++ app/src/main/java/com/example/team25/data/network/dto/HospitalDto.kt
@@ -1,4 +1,4 @@
-package com.example.team25.ui.reservation.network.dto
+package com.example.team25.data.network.dto
 
 import com.google.gson.annotations.SerializedName
