+++ app/src/main/java/com/example/team25/data/network/services/RemoteSearchHospitalService.kt
@@ -1,9 +1,9 @@
-package com.example.team25.ui.reservation.services
+package com.example.team25.data.network.services
 
 import com.example.team25.BuildConfig
 import com.example.team25.domain.HospitalDomain
 import com.example.team25.ui.reservation.interfaces.SearchHospitalService
-import com.example.team25.ui.reservation.network.KakaoApi
+import com.example.team25.data.network.KakaoApi
 import javax.inject.Inject
 
 class RemoteSearchHospitalService
